requirements.txt
absl-py==0.13.0
alabaster==0.7.12
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
docutils==0.18
entrypoints==0.3
Fiona==1.8.20
flatbuffers==1.12
future==0.18.2
gast==0.4.0
geopandas==0.9.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.39.0
h5py==3.1.0
idna==2.6
imagesize==1.2.0
importlib-metadata==4.6.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
livereload==2.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.2
mistune==0.8.4
munch==2.5.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.6
notebook==6.4.6
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.1.3
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
pockets==0.9.1
port-for==0.3.1
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.17.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
Pygments==2.10.0
pygobject==3.26.1
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.18.0
pytest==6.1.2
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
SecretStorage==2.3.1
Send2Trash==1.8.0
Shapely==1.8.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==3.1.2
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.6
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xlrd==2.0.1
zipp==3.5.0
